+++ packages/harpparts/src/access-parts/constants/tunings.ts
@@ -5,7 +5,9 @@ import {
   NATURAL_MINOR,
   HARMONIC_MINOR,
   MELODY_MAKER,
+  EASY_DIATONIC,
   WILDE,
+  WILDE_MINOR,
   CIRCULAR,
   ORCHESTRA_S,
   POWER_BENDER,
@@ -45,7 +47,9 @@ orderedTunings.set(TuningIds.Country, COUNTRY)
 orderedTunings.set(TuningIds.NaturalMinor, NATURAL_MINOR)
 orderedTunings.set(TuningIds.HarmonicMinor, HARMONIC_MINOR)
 orderedTunings.set(TuningIds.MelodyMaker, MELODY_MAKER)
+orderedTunings.set(TuningIds.EasyDiatonic, EASY_DIATONIC)
 orderedTunings.set(TuningIds.Wilde, WILDE)
+orderedTunings.set(TuningIds.WildeMinor, WILDE_MINOR)
 orderedTunings.set(TuningIds.Circular, CIRCULAR)
 orderedTunings.set(TuningIds.OrchestraS, ORCHESTRA_S)
 orderedTunings.set(TuningIds.PowerBender, POWER_BENDER)